Fixing Expo for phone and Web (#47)
Not tested on Android but should work. Works on ios and web, access
control for images and camera works, and buttons have been updated to be
compatible both ios and web with more customizability with stylesheet
integration. Updated and downgraded some dependencies for compatibility.
npm audit fix DOES NOT work! current attempts to fix the vulnerabilities
will break expo.