Clear forms (#133)
# Clear forms: Resetting review and facility form on submit (#92)

# Checklist

- [x] Have I clearly explained the issue or feature request that this PR
addresses?
- [x] Have I provided where new implementations for the PR have been
made?
- [x] Did I highlight any areas that may need special attention during
the review?
- [x] If applicable, did I specify any setup requirements or
dependencies for testing?
- [x] Have I disclosed any known issues or limitations related to this
PR? Did I explain why these limitations exist and how they might be
addressed in the future?
- [x] Have I provided enough context for someone unfamiliar with the
issue or the codebase to understand the changes?
- [x] Have I assigned appropriate reviewers who are knowledgeable about
the changes? Did I consider including team members who might be affected
or interested in the changes?

# Description
- Issue #92 
- Facility form and review form clears text/images/stars/etc on submit
- Both also reset scroll position to the top of the screen
- Review form now also clears tags

## Expectations
  - Facility Review and Form have been modified
- Both now have a ResetForm function that is called on submit that
clears all input.

## Known issues:
- When an image is uploaded and no other fields are inputted, then the
photos are not reset on submit. The photos are deleted, but the
component does not refresh until the add photo button is pressed.
- While an issue, I propose another PR should be submitted to fix this
later as main functionality is implemented and the issue doesn't greatly
impact/break anything (i.e. is purely visual).