Add workflow badges to README (#134)
<!---

PR Title: [Brief Description] (no issue number)

PR Description:
- Provide a summary of the issue:
- Include the issue number and a direct link to the issue in the
repository
  - Summarize the issue that this PR is addressing
  - Include the problem or feature request and why it's important
- Provide a set of expectations:
  - Include a detailed overview of the changes made in this PR
  - Include information about new features, bug fixes, or improvements
- Explain how these changes affect the existing functionality of the
application
  - Highlight any areas that need special attention during the review
  - List the steps reviewers should take to test the changes
- Include any setup requirements, specific scenarios to test, and
expected outcomes
- Additional Comments/ Notes:
- Mention any known issues or limitations of this PR that reviewers
should be aware of
- If applicable, outline any future work that will follow this PR. This
helps reviewers
    understand the broader context of the changes.

PR Reviewers: Assign individuals who can thoughtfully review your PR.
Consider who is familiar with the codebase and the specific areas of
development related to this PR

Development (sidebar): List any issues that this PR is associated with

--->

# Checklist

- [x] Have I clearly explained the issue or feature request that this PR
addresses?
- [x] Have I provided where new implementations for the PR have been
made?
- [x] Did I highlight any areas that may need special attention during
the review?
- [x] If applicable, did I specify any setup requirements or
dependencies for testing?
- [x] Have I disclosed any known issues or limitations related to this
PR? Did I explain why these limitations exist and how they might be
addressed in the future?
- [x] Have I provided enough context for someone unfamiliar with the
issue or the codebase to understand the changes?
- [x] Have I assigned appropriate reviewers who are knowledgeable about
the changes? Did I consider including team members who might be affected
or interested in the changes?

# Description

## Summary

Completes frontend portion of Green-Kittens/bathroom-finder#66

Professor K said we should do this. I think it's a grading requirement.

## Expectations

- Build badges will appear on the repo's overall README
- They will reflect the result of the workflows being ran on whatever is
the current latest commit on `main`
- The Compile, Lint, and Format workflows will now run on every push to
main to facilitate this
- This is admittedly inefficient since we will also have ran the
workflows during PR, but GitHub Actions does not cache those results so
this is how we have to do it

## Additional Comments/Notes

- badges will show as "no status" until this PR is merged, as the
workflows will not run on `main` (as is required, see above) until the
`yml` file changes are merged

Co-authored-by: Lindsay Landstad <71728910+lindlands@users.noreply.github.com>