fix: polish description on facility profile (#131)
## Summary
Closes #76

The previous implementation of the "view more text" component on the
facility profile had a laggy animation which made it feel like something
was wrong when you expanded it. I used an external library's component
instead, doing away with the animation.

This is important so that our users are confident in our app.
Eliminating that laggy animation takes away the fear that there are
performance issues.

## Expectations
- There should only be changes to the description section on the
facility profile, everything else should remain the same.
- When clicking the "view more"/"view less" button, the text should just
jump between the two sizes, rather than animating.
- There are some minor styling changes, since I'm using a completely
different implementation of the component. These styling changes are
intended to be trivial. If you think that they are more than trivial,
please comment as such.
- Description text is still placeholder. Getting the description from
the controller is outside the scope of this PR.

## Additional Comments/Notes
- To view the facility profile description
    1. Load the app at the home/map screen
    2. Zoom out until you see one of the location markers on the map
3. Click the location marker to open that location's facility profile
    4. Look just under the "Add Review" button on that screen.
5. The paragraph should be 3 lines, end in a `...`, and have a "view
more" button upon loading the facility profile page.
6. Upon clicking the "view more" button, the paragraph should extend to
full length (number of lines depends on length of content, no `...`) and
the "view more" button should change to a "view less" button.
- Enjoying the PR template.Thanks @meganfung38!!